更改fegin位置

+++ cjyc-common-system/src/main/java/com/cjyc/common/system/service/ICsAdminService.java
@@ -22,5 +22,5 @@ public interface ICsAdminService {
      */
     List<Admin> getListByStoreId(Long storeId);
 
-    CacheAdminVo getCacheData(Long userId, Long roleId);
+    CacheAdminVo getCacheData(Long userId, Integer roleId);
 }
